Add LinkButton
The HTML5 spec says that a button must not have any "interactive content
descendant". This includes <a>-tags. This commit introduces a new tag,
LinkButton, which has the same styling as a Button, but is a <a> instead
of <button>.

LinkButton does not support `action` since it's supposed to only be used
for actual links.